agent/requirements.txt
# Existing dependencies
langchain>=0.1.0
langchain-google-genai>=0.0.5
python-dotenv>=1.0.0
textual>=0.47.1
rich>=13.7.0
langgraph>=0.0.20
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.108.0

# New: Redis for eval queue
redis>=5.0.0

# New: HTTP client for eval server
httpx>=0.25.0
